only expose services that need public access
Also the exposed services are going to use a LoadBalancer instead of
NodePort. Both expose the service but the Load Balancer option is better
when on the cloud, since the cloud provider will schedule a Load
Balancer with a public IP.

Though it would be better to have a Ingress file to do smart routing, so
we do not pay for multiple Load Balancers and external IPs.

Signed-off-by: Peter Bom Jakobsen <pejak1990@gmail.com>